+++ vitest.global-setup.ts
@@ -13,9 +13,7 @@ let mongod: MongoMemoryServer;
 export async function setup(project: TestProject) {
   mongod = await MongoMemoryServer.create({
     instance: {
-      // https://github.com/nodkz/mongodb-memory-server/issues/78
-      // storageEngine: 'wiredTiger'
-      storageEngine: 'ephemeralForTest'
+      storageEngine: 'wiredTiger'
     }
   });
   // @ts-ignore