+++ changelog.md
@@ -15,6 +15,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 - Use the document’s `_id` if no slug can be created (when slugifying the string doesn’t give any characters)
 - Update `limax` to version 4.2.3
+- Compatibility with Mongoose 9
 
 ## [6.1.1] – 2024-04-30